refactor: Rename Robolectric test classes to Instrumentation
